\name{RTajo}
\alias{RTajo}
\title{A Package to interface R and Tajo}

\description{
RTajo provides both parallel interactive and batch processing capabilities connecting to the Tajo
}

\author{
wonkuk Yang \email{gigsda@gmail.com}
}

\details{
\tabular{ll}{
Package: \tab RTajo\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2015-04-25\cr
License: \tab file LICENSE\cr
}
}

\section{Installation}{
}

\references{
http://tajo.apache.org
}
\keyword{ Impala }
\keyword{ Installation }
\keyword{ JDBC }
\keyword{ drivers }
